+++ support-frontend/ts-error-history.csv
@@ -128,4 +128,6 @@
 1645113025000,363,e69ca791a7f8e569e7013521d69f497e4a362109
 1645453585000,363,969ac3f83873e233199d750d0837e7ada9f23fe7
 1645453715000,363,194633e0a7aa1611e82bd9f06cbcd0233049ca6c
-1645454500000,363,HEAD
+1645455801000,363,633836d6dbbd310d3a1b44e37274525edb23479f
+1645455923000,363,2f8396f01ac24f9993812e677e2db0053454d318
+1645710406000,358,HEAD
