+++ support-frontend/typescript-errors.json
@@ -3,7 +3,7 @@
         "byCode": {
             "TS1208": 1,
             "TS2305": 12,
-            "TS2322": 76,
+            "TS2322": 73,
             "TS2578": 5,
             "TS2339": 15,
             "TS2345": 54,
@@ -14,8 +14,8 @@
             "TS2769": 14,
             "TS2314": 4,
             "TS2304": 5,
-            "TS2741": 4,
-            "TS2739": 13,
+            "TS2741": 3,
+            "TS2739": 12,
             "TS2532": 4,
             "TS2556": 3,
             "TS2554": 27,
@@ -71,7 +71,6 @@
             "assets/pages/contributions-landing/components/AmazonPay/AmazonPayLoginButton.tsx": 2,
             "assets/pages/contributions-landing/components/SepaForm.tsx": 7,
             "assets/pages/contributions-landing/components/StripeCardForm/StripeCardFormContainer.tsx": 1,
-            "assets/pages/contributions-landing/contributionsLandingReducer.ts": 5,
             "assets/pages/contributions-landing/setUserStateActions.ts": 3,
             "assets/pages/digital-subscription-checkout/components/digitalCheckoutForm.test.tsx": 9,
             "assets/pages/digital-subscription-checkout/components/digitalCheckoutForm.tsx": 4,
@@ -393,18 +392,6 @@
         {
             "path": "assets/pages/contributions-landing/components/StripeCardForm/StripeCardFormContainer.tsx",
             "message": "Type '{ stripeKey: string; isTestUser: boolean; }' is not assignable to type 'IntrinsicAttributes & Omit<{ contributionType: keyof RegularContributionTypeMap<null> | \"ONE_OFF\"; checkoutFormHasBeenSubmitted: boolean; ... 9 more ...; isTestUser: boolean; } & { ...; } & { ...; }, \"csrf\" | ... 19 more ... | \"setStripeRecurringRecaptchaVerified\">'."
-        },
-        {
-            "path": "assets/pages/contributions-landing/contributionsLandingReducer.ts",
-            "message": "Type '{ Stripe: null; AmazonPay: null; }' is not assignable to type '{ Stripe: StripeHandler | null; }'."
-        },
-        {
-            "path": "assets/pages/contributions-landing/contributionsLandingReducer.ts",
-            "message": "Type '{ hasBegunLoading: false; amazonPayLibrary: { amazonLoginObject: null; amazonPaymentsObject: null; }; loginButtonReady: false; walletIsStale: false; orderReferenceId: null; paymentSelected: false; hasAccessToken: false; fatalError: false; amazonBillingAgreementConsentStatus: false; }' is not assignable to type 'AmazonPayData'."
-        },
-        {
-            "path": "assets/pages/contributions-landing/contributionsLandingReducer.ts",
-            "message": "Type '{ contributionType: keyof RegularContributionTypeMap<null> | \"ONE_OFF\"; showOtherAmount: boolean; formData: { otherAmounts: OtherAmounts; ... 5 more ...; email: string | null; }; ... 17 more ...; oneOffRecaptchaToken: string | null; }' is not assignable to type 'FormState'."
         },
         {
             "path": "assets/pages/contributions-landing/setUserStateActions.ts",
@@ -1241,10 +1228,6 @@
             "message": "Property 'label' is missing in type '{ onKeyPress: (e: KeyboardEvent<HTMLInputElement>) => void; css: SerializedStyles; name: string; width: 10; }' but required in type 'TextInputProps'."
         },
         {
-            "path": "assets/pages/contributions-landing/contributionsLandingReducer.ts",
-            "message": "Property 'Widgets' is missing in type 'Record<string, any>' but required in type 'AmazonPaymentsObject'."
-        },
-        {
             "path": "assets/pages/digital-subscription-checkout/components/endSummary/endSummarySelector.ts",
             "message": "Property 'isPatron' is missing in type '{ priceDescription: string; promotion: string; orderIsAGift: boolean; digitalGiftBillingPeriod: \"Annual\" | \"Quarterly\"; price: string; }' but required in type 'EndSummaryProps'."
         },
@@ -1297,10 +1280,6 @@
         {
             "path": "assets/helpers/internationalisation/localCurrencyCountry.ts",
             "message": "Type '{ ONE_OFF: { min: number; minInWords: string; max: number; maxInWords: string; default: number; }; }' is missing the following properties from type 'Config': MONTHLY, ANNUAL"
-        },
-        {
-            "path": "assets/pages/contributions-landing/contributionsLandingReducer.ts",
-            "message": "Type '{}' is missing the following properties from type 'SelectedAmounts': MONTHLY, ANNUAL, ONE_OFF"
         },
         {
             "path": "assets/pages/paper-subscription-checkout/components/thankYou.tsx",
